Move action and mutation types to constants directory
