+++ src/components/HeroSection.vue
@@ -59,7 +59,7 @@
 </template>
 
 <script>
-import { SET_QUERY } from '~/store-modules/mutation-types'
+import { SET_QUERY } from '~/constants/mutation-types'
 import { filtersToQueryData } from '~/utils/search-query-transform'
 import { mapState } from 'vuex'
 import { ALL_MEDIA } from '~/constants/media'
