+++ src/components/ImageDetails/SocialShareButtons.vue
@@ -36,7 +36,7 @@
 </template>
 
 <script>
-import { SOCIAL_MEDIA_SHARE } from '~/store-modules/action-types'
+import { SOCIAL_MEDIA_SHARE } from '~/constants/action-types'
 import {
   SEND_DETAIL_PAGE_EVENT,
   DETAIL_PAGE_EVENTS,
