+++ src/components/NavSection.vue
@@ -125,7 +125,7 @@
 </template>
 
 <script>
-import { SET_QUERY } from '~/store-modules/mutation-types'
+import { SET_QUERY } from '~/constants/mutation-types'
 import Dropdown from '~/components/Dropdown'
 
 export default {
