+++ src/components/PhotoTags.vue
@@ -19,7 +19,7 @@
 </template>
 
 <script>
-import { SET_QUERY } from '~/store-modules/mutation-types'
+import { SET_QUERY } from '~/constants/mutation-types'
 
 export default {
   name: 'PhotoTags',
