+++ src/components/SearchGridForm.vue
@@ -60,7 +60,7 @@
 </template>
 
 <script>
-import { SET_FILTER_IS_VISIBLE } from '~/store-modules/mutation-types'
+import { SET_FILTER_IS_VISIBLE } from '~/constants/mutation-types'
 import { queryStringToSearchType } from '~/utils/search-query-transform'
 import { VIDEO } from '~/constants/media'
