+++ src/pages/search/index.vue
@@ -8,7 +8,7 @@
 </template>
 
 <script>
-import { UPDATE_SEARCH_TYPE } from '~/store-modules/action-types'
+import { UPDATE_SEARCH_TYPE } from '~/constants/action-types'
 import { ALL_MEDIA } from '~/constants/media'
 
 export default {
