+++ src/pages/search/video.vue
@@ -8,7 +8,7 @@
 </template>
 
 <script>
-import { UPDATE_SEARCH_TYPE } from '~/store-modules/action-types'
+import { UPDATE_SEARCH_TYPE } from '~/constants/action-types'
 import { VIDEO } from '~/constants/media'
 
 export default {
