+++ src/store-modules/report-content-store.js
@@ -4,8 +4,8 @@ import {
   REPORT_FORM_CLOSED,
   REPORT_ERROR,
   BACK_TO_REPORT_START,
-} from './mutation-types'
-import { SEND_CONTENT_REPORT } from './action-types'
+} from '../constants/mutation-types'
+import { SEND_CONTENT_REPORT } from '../constants/action-types'
 
 const state = {
   isReportFormVisible: false,
