+++ test/unit/specs/components/photo-tags.spec.js
@@ -1,5 +1,5 @@
 import PhotoTags from '~/components/PhotoTags'
-import { SET_QUERY } from '~/store-modules/mutation-types'
+import { SET_QUERY } from '~/constants/mutation-types'
 import render from '../../test-utils/render'
 
 describe('PhotoTags', () => {
